Add files via upload
This is a comprehensive paper with code of how the Hizen formula can develop AGI. This is a test case only paper and I NEED help in refinement from the community that is cloneing this Git Hub. If you feel this may work please reach out to me to collaborate as we are 86 seconds to Death or Dance as a species.